+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Core/Extensions/FloaterCollection.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Unity.Burst;
 using Unity.Mathematics;
 using UnityEngine;
 
@@ -6,16 +7,19 @@ namespace ChoyUtilities {
     
     public static partial class HelperCollection {
 
+        [BurstCompile]
         public static float3 PositionFromTransform(this Floater data) {
             return data.Length < 9 ? float3.zero : new float3(data[0], data[1], data[2]);
         }
         
+        [BurstCompile]
         public static quaternion RotationFromTransform(this Floater data) {
             if (data.Length < 9) return quaternion.identity;
             var euler = new Vector3(data[3], data[4], data[5]);
             return quaternion.Euler(euler);
         }
         
+        [BurstCompile]
         public static float3 ScaleFromTransform(this Floater data) {
             return data.Length < 9 ? new float3(1, 1, 1) : new float3(data[6], data[7], data[8]);
         }