+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Core/Extensions/RawSetCollection.cs
@@ -1,9 +1,11 @@
-﻿using Unity.Mathematics;
+﻿using Unity.Burst;
+using Unity.Mathematics;
 using UnityEngine;
 
 namespace ChoyUtilities {
     public static partial class HelperCollection {
         
+        [BurstCompile]
         public static Transform RawSet(this Transform obj, RawSet<float3> set) {
             if (!set.IsCreated || set.Length < 3) throw new RawSetException("Length of RawSet need to more than 3");
             var pos = set[0];