+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Enums/EMotion.cs
@@ -1,0 +1,9 @@
+﻿namespace ChoyUtilities {
+    
+    public enum EMotion : byte{
+        Linear,
+        EaseIn,
+        EaseOut,
+        EaseInOut
+    }
+}