+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Motion/Utils/MotionUtils.cs
@@ -1,0 +1,6 @@
+﻿namespace ChoyUtilities {
+    
+    public static partial class HelperCollection {
+        
+    }
+}